Started work on work log controller.

+++ RandomFruit/app/database/migrations/2014_04_18_150116_create_work_logs_table.php
@@ -22,7 +22,7 @@ class CreateWorkLogsTable extends Migration {
 			$table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
 			$table->integer('week_id')->unsigned();
 			$table->foreign('week_id')->references('id')->on('weeks')->onDelete('cascade');
-			$table->integer('value');
+			$table->decimal('value', 6, 1);
 			$table->timestamps();
 		});
 	}
